drf_project/requirements.txt
asgiref==3.8.1
attrs==23.2.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
defusedxml==0.8.0rc2
Django==5.0.6
django-crontab==0.7.1
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
h11==0.14.0
idna==3.7
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==24.1
pycparser==2.22
PyJWT==2.8.0
PySocks==1.7.1
python-dotenv==1.0.1
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
selenium==4.22.0
sniffio==1.3.1
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sortedcontainers==2.4.0
sqlparse==0.5.0
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
webdriver-manager==4.0.1
websocket-client==1.8.0
wsproto==1.2.0
